+++ lib/internal/console/constructor.js
@@ -1,25 +1,7 @@
-// Copyright Joyent, Inc. and other Node contributors.
-//
-// Permission is hereby granted, free of charge, to any person obtaining a
-// copy of this software and associated documentation files (the
-// "Software"), to deal in the Software without restriction, including
-// without limitation the rights to use, copy, modify, merge, publish,
-// distribute, sublicense, and/or sell copies of the Software, and to permit
-// persons to whom the Software is furnished to do so, subject to the
-// following conditions:
-//
-// The above copyright notice and this permission notice shall be included
-// in all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
-// OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
-// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN
-// NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
-// DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
-// OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE
-// USE OR OTHER DEALINGS IN THE SOFTWARE.
-
 'use strict';
+
+// The Console constructor is not actually used to construct the global
+// console. It's exported for backwards compatibility.
 
 const { trace } = internalBinding('trace_events');
 const {
@@ -72,8 +54,6 @@ const kBindStreamsLazy = Symbol('kBindStreamsLazy');
 const kUseStdout = Symbol('kUseStdout');
 const kUseStderr = Symbol('kUseStderr');
 
-// This constructor is not used to construct the global console.
-// It's exported for backwards compatibility.
 function Console(options /* or: stdout, stderr, ignoreErrors = true */) {
   // We have to test new.target here to see if this function is called
   // with new, because we need to define a custom instanceof to accommodate
@@ -531,41 +511,8 @@ Console.prototype.table = function(tabularData, properties) {
 
 function noop() {}
 
-// See https://console.spec.whatwg.org/#console-namespace
-// > For historical web-compatibility reasons, the namespace object
-// > for console must have as its [[Prototype]] an empty object,
-// > created as if by ObjectCreate(%ObjectPrototype%),
-// > instead of %ObjectPrototype%.
-
-// Since in Node.js, the Console constructor has been exposed through
-// require('console'), we need to keep the Console constructor but
-// we cannot actually use `new Console` to construct the global console.
-// Therefore, the console.Console.prototype is not
-// in the global console prototype chain anymore.
-
-// TODO(joyeecheung):
-// - Move the Console constructor into internal/console.js
-// - Move the global console creation code along with the inspector console
-//   wrapping code in internal/bootstrap/node.js into a separate file.
-// - Make this file a simple re-export of those two files.
-const globalConsole = Object.create({});
-
-// Since Console is not on the prototype chain of the global console,
-// the symbol properties on Console.prototype have to be looked up from
-// the global console itself. In addition, we need to make the global
-// console a namespace by binding the console methods directly onto
-// the global console with the receiver fixed.
-for (const prop of Reflect.ownKeys(Console.prototype)) {
-  if (prop === 'constructor') { continue; }
-  const desc = Reflect.getOwnPropertyDescriptor(Console.prototype, prop);
-  if (typeof desc.value === 'function') { // fix the receiver
-    desc.value = desc.value.bind(globalConsole);
-  }
-  Reflect.defineProperty(globalConsole, prop, desc);
-}
-
-globalConsole[kBindStreamsLazy](process);
-globalConsole[kBindProperties](true, 'auto');
-
-module.exports = globalConsole;
-module.exports.Console = Console;
+module.exports = {
+  Console,
+  kBindStreamsLazy,
+  kBindProperties
+};
